backend/requirements.txt
fastapi==0.109.0
fastapi-utils==0.2.1
uvicorn==0.27.0
sqlalchemy==2.0.25
mysql-connector-python==8.3.0
